Propagate verify as a parameter on request
It seems like propagating `verify` as a session parameter does not
work in all cases. In those cases when it does not work, it appears that
propagating it as a keyword argument in the request itself would be
better.

Additionally, this PR does a small refactor on how the `kwargs`
dictionary is constructed, by separating it into a few functions so the
intention of the code is clearer.