[fix] Correctly apply the SSL configuration (#62)
[Fix] Correctly apply the SSL configuration
<!-- PR title should start with '[fix]', '[improvement]' or '[break]' if this PR would cause a patch, minor or major SemVer bump. Omit the prefix if this PR doesn't warrant a standalone release. -->

## Before this PR
<!-- Describe the problem you encountered with the current state of the world (or link to an issue) and why it's important to fix now. -->
It seems like propagating `verify` as a session parameter does not
work in all cases. In those cases when it does not work, it appears that
propagating it as a keyword argument in the request itself works.

## After this PR
<!-- Describe at a high-level why this approach is better. -->
The `verify` parameter is now propagated on each individual request. 

Additionally, this PR does a small refactor for how the `kwargs`
dictionary is constructed, by separating it into a few functions so the
intention of the code is clearer.

<!-- Reference any existing GitHub issues, e.g. 'fixes #000' or 'relevant to #000' -->